requirements.txt
requests==2.26.0
discord.py==1.7.3
beautifulsoup4==4.10.0
discord==1.7.3
googletrans==4.0.0rc1
nest_asyncio==1.5.1
python-dotenv==0.19.0
regex==2021.9.30
scrapy==2.5.0
pandas==1.3.4
matplotlib==3.4.3
seaborn==0.11.2
Flask==1.1.4
